_camera_roi.py
